+++ src/Exceptions/NotValidEloquentException.php
@@ -1,0 +1,9 @@
+<?php
+namespace MAbadir\ElasticLaravel\Exceptions;
+
+use Exception;
+
+class NotValidEloquentException extends Exception
+{
+
+}